Remove venv do versionamento

+++ app/services/cidade_service.py
@@ -1,45 +1,77 @@
 from sqlalchemy.orm import Session
 from fastapi import HTTPException
-from app.schemas.cidade_schema import ResponseCidadesSchema, CidadesSchema
+from app.schemas.cidade_schema import CidadeCreate, CidadeOut, CidadeUpdate
 from app.models.cidade_models import Cidade
-from typing import Optional
+from app.services.clube_service import buscar_clube_por_cidade_id
+from app.services.estadio_service import buscar_estadio_por_cidade_id
+from typing import Optional, List
 import re
 
 
-def listar_todas_cidades(session: Session) -> ResponseCidadesSchema:
+def listar_todas_cidades(nome: Optional[str], uf: Optional[str], session: Session) -> List[CidadeOut]:
     """
     Serviço para listar todas as cidades no banco de dados.
 
     Args:
+        nome (Optional[str]): Nome da cidade a ser filtrada (opcional).
+        uf (Optional[str]): UF da cidade a ser filtrada (opcional).
         session (Session): Sessão ativa do SQLAlchemy para conectar ao banco.
 
     Raises:
         HTTPException: Quando não há cidades cadastradas no banco de dados.
 
     Returns:
-        ResponseCidadesSchema: Contém uma lista de cidades no formato esperado na API.
+        CidadeCreate: Contém uma lista de cidades no formato esperado na API.
     """
     # Obtemos todas as cidades do banco de dados
-    cidades = session.query(Cidade).order_by(Cidade.__table__.c.cid_nome).all()
+    query = session.query(Cidade)
+    if nome:
+        nome = (
+            nome.strip()
+            .replace("%", "\\%")
+            .replace("_", "\\_")
+        )
+
+        query = query.filter(
+            Cidade.cid_nome.ilike(
+                f"%{nome}%",
+                escape="\\"
+            )
+        )
+
+    if uf:
+        uf = uf.strip().upper()
+
+        query = query.filter(
+            Cidade.cid_uf == uf
+        )
+
+    query = query.order_by(
+        Cidade.cid_nome
+    ).all()
 
     # Validação quando não houver cidades cadastradas
-    if not cidades:
+    if not query:
         raise HTTPException(status_code=404, detail="Não há cidades cadastradas.")
     
-    # Convertendo os objetos Cidade para o formato do schema
-    cidades_schema = [CidadesSchema(**cidade.as_dict()) for cidade in cidades]
+    return [
+        CidadeOut(
+                cid_id = c.cid_id,
+                cid_nome = c.cid_nome,
+                cid_uf = c.cid_uf,
+        )
+        for c in query
+    ]
 
-    return ResponseCidadesSchema(cidades=cidades_schema)
-
-def criar_cidade(cidade: CidadesSchema, session: Session) -> CidadesSchema:
+def criar_cidade(cidade: CidadeCreate, session: Session) -> CidadeCreate:
     """Criar registro na tabela de cidade
 
     Args:
         session (Session): Sessão ativa do SQLAlchemy para conectar ao banco.
-        cidade (CidadesSchema): Dados da cidade a ser criada.
+        cidade (CidadeCreate): Dados da cidade a ser criada.
 
     Returns:
-        CidadesSchema: Dados da cidade criada.
+        CidadeCreate: Dados da cidade criada.
     """
     if buscar_cidade_nome(False, cidade.cid_nome, session):  # Verifica se a cidade já existe
         raise HTTPException(status_code=404, detail=f"Cidade {cidade.cid_nome.title()} já cadastrada.")
@@ -52,24 +84,23 @@ def criar_cidade(cidade: CidadesSchema, session: Session) -> CidadesSchema:
     session.add(nova_cidade)
     session.commit()
     session.refresh(nova_cidade)
-    return CidadesSchema(**nova_cidade.as_dict())
+    return CidadeCreate(**nova_cidade.as_dict())
 
 
-def atualizar_cidade(cidade_id: int, cidade_atualizada: CidadesSchema, session: Session) -> CidadesSchema:
+def atualizar_cidade(cidade_atualizada: CidadeUpdate, session: Session) -> CidadeUpdate:
     """Atualizar registro na tabela de cidade
 
     Args:
-        cidade_id (int): ID da cidade a ser atualizada.
-        cidade_atualizada (CidadesSchema): Dados atualizados da cidade.
+        cidade_atualizada (CidadeUpdate): Dados atualizados da cidade.
         session (Session): Sessão ativa do SQLAlchemy para conectar ao banco.
 
     Raises:
         HTTPException: Lançada se a cidade não for encontrada no banco de dados.
 
     Returns:
-        CidadesSchema: Dados da cidade atualizada.
+        CidadeUpdate: Dados da cidade atualizada.
     """
-    cidade_db = session.query(Cidade).filter(Cidade.cid_id == cidade_id).first()
+    cidade_db = session.query(Cidade).filter(Cidade.cid_id == cidade_atualizada.cid_id).first()
 
     if not cidade_db:
         raise HTTPException(status_code=404, detail="Cidade não encontrada.")
@@ -79,7 +110,7 @@ def atualizar_cidade(cidade_id: int, cidade_atualizada: CidadesSchema, session: 
         nome_normalizado = re.sub(r'\s+', ' ', cidade_atualizada.cid_nome.strip()).title()
 
         cidade_existente = buscar_cidade_nome(False, nome_normalizado, session)
-        if cidade_existente and cidade_existente.cid_nome == nome_normalizado:
+        if cidade_existente and cidade_existente.cid_nome == nome_normalizado and cidade_existente.cid_id != cidade_atualizada.cid_id:
             raise HTTPException(status_code=404, detail=f"Cidade {cidade_atualizada.cid_nome.strip().title()} já cadastrada.")
         
         cidade_db.cid_nome = nome_normalizado
@@ -92,7 +123,7 @@ def atualizar_cidade(cidade_id: int, cidade_atualizada: CidadesSchema, session: 
         session.commit()
         session.refresh(cidade_db)
 
-    return CidadesSchema(**cidade_db.as_dict())
+    return CidadeUpdate(**cidade_db.as_dict())
 
 def deletar_cidade(cidade_id: int, session: Session):
     """Deletar registro na tabela de cidade
@@ -108,16 +139,25 @@ def deletar_cidade(cidade_id: int, session: Session):
     if not cidade_db:
         raise HTTPException(status_code=404, detail="Cidade não encontrada.")
     
+    clube_associado_cidade = buscar_clube_por_cidade_id(False, cidade_id, session)
+    if clube_associado_cidade:
+        raise HTTPException(status_code=404, detail=f"Não é possível excluir a cidade pois existe clube associado a ela: {clube_associado_cidade.clu_nome}.")
+
+    estadio_associado_cidade = buscar_estadio_por_cidade_id(False, cidade_id, session)
+    if estadio_associado_cidade:
+        raise HTTPException(status_code=404, detail=f"Não é possível excluir a cidade pois existe estádio associado a ela: {estadio_associado_cidade.est_nome}.")
+    
     session.delete(cidade_db)
     session.commit()
     return "Cidade excluida com sucesso"
 
 
-def listar_cidades_paginadas(nome: Optional[str], pagina: int, tamanho_pagina: int, session: Session) -> ResponseCidadesSchema:
+def listar_cidades_paginadas(nome: Optional[str], uf: Optional[str], pagina: int, tamanho_pagina: int, session: Session) -> List[CidadeOut]:
     """Listar as cidades pelo nome da cidade (opcional) com paginação
 
     Args:
         nome (Optional[str]): Nome da cidade a ser filtrada (opcional).
+        uf (Optional[str]): UF da cidade a ser filtrada (opcional).
         pagina (int): Número da página a ser retornada.
         tamanho_pagina (int): Tamanho da página a ser retornada.
         session (Session): Sessão ativa do SQLAlchemy para conectar ao banco.
@@ -126,11 +166,14 @@ def listar_cidades_paginadas(nome: Optional[str], pagina: int, tamanho_pagina: i
         HTTPException: Lançada se nenhuma cidade for encontrada.
 
     Returns:
-        ResponseCidadesSchema: Lista de cidades no formato esperado na API.
+        List[CidadeCreate]: Lista de cidades no formato esperado na API.
     """
     query = session.query(Cidade)
     if nome:
         query = query.filter(Cidade.__table__.c.cid_nome.ilike(f"%{nome}%"))
+
+    if uf:
+        query = query.filter(Cidade.__table__.c.cid_uf.ilike(f"%{uf.upper()}%"))
 
     query = query.order_by(Cidade.__table__.c.cid_nome)
         
@@ -141,12 +184,17 @@ def listar_cidades_paginadas(nome: Optional[str], pagina: int, tamanho_pagina: i
     if not cidades:
         raise HTTPException(status_code=404, detail="Nenhuma cidade encontrada.")
     
-    cidades_schema = [CidadesSchema(**cidade.as_dict()) for cidade in cidades]
-    
-    return ResponseCidadesSchema(cidades=cidades_schema)
+    return [
+        CidadeOut(
+            cid_id = c.cid_id,
+            cid_nome = c.cid_nome,
+            cid_uf = c.cid_uf,
+        )
+        for c in cidades
+    ]
 
 
-def buscar_cidade_nome(retorna_exception: bool, nome: str, session: Session) -> Optional[CidadesSchema]:
+def buscar_cidade_nome(retorna_exception: bool, nome: str, session: Session) -> Optional[CidadeOut]:
     """
     Busca uma cidade pelo nome no banco de dados.
 
@@ -159,7 +207,7 @@ def buscar_cidade_nome(retorna_exception: bool, nome: str, session: Session) -> 
         HTTPException: Caso a cidade não seja encontrada.
 
     Returns:
-        Optional[CidadesSchema]: Representação da cidade encontrada ou None se não encontrada.
+        Optional[CidadeCreate]: Representação da cidade encontrada ou None se não encontrada.
     """
     # Busca pela cidade no banco de dados (ignora case com ilike)
     cidade = session.query(Cidade).filter(Cidade.__table__.c.cid_nome.ilike(f"%{nome}%")).first()
@@ -168,9 +216,9 @@ def buscar_cidade_nome(retorna_exception: bool, nome: str, session: Session) -> 
         if retorna_exception:
             raise HTTPException(status_code=404, detail=f"Cidade com nome '{nome}' não encontrada.")
 
-    return CidadesSchema(**cidade.as_dict()) if cidade else None
+    return CidadeOut(**cidade.as_dict()) if cidade else None
 
-def buscar_cidade_id(retorna_exception: bool, cidade_id: int, session: Session) -> Optional[CidadesSchema]:
+def buscar_cidade_id(retorna_exception: bool, cidade_id: int, session: Session) -> Optional[CidadeOut]:
     """
     Busca uma cidade pelo ID no banco de dados.
 
@@ -183,7 +231,7 @@ def buscar_cidade_id(retorna_exception: bool, cidade_id: int, session: Session) 
         HTTPException: Caso a cidade não seja encontrada.
 
     Returns:
-        Optional[CidadesSchema]: Representação da cidade encontrada ou None se não encontrada.
+        Optional[CidadeOut]: Representação da cidade encontrada ou None se não encontrada.
     """
     # Busca pela cidade no banco de dados
     cidade = session.query(Cidade).filter(Cidade.cid_id == cidade_id).first()
@@ -192,9 +240,9 @@ def buscar_cidade_id(retorna_exception: bool, cidade_id: int, session: Session) 
         if retorna_exception:
             raise HTTPException(status_code=404, detail=f"Cidade com ID '{cidade_id}' não encontrada.")
 
-    return CidadesSchema(**cidade.as_dict()) if cidade else None
+    return CidadeOut(**cidade.as_dict()) if cidade else None
 
-def buscar_cidade_uf(retorna_exception: bool, uf: str, session: Session) -> Optional[CidadesSchema]:
+def buscar_cidade_uf(retorna_exception: bool, uf: str, session: Session) -> Optional[CidadeOut]:
     """
     Busca uma cidade pela UF no banco de dados.
 
@@ -207,7 +255,7 @@ def buscar_cidade_uf(retorna_exception: bool, uf: str, session: Session) -> Opti
         HTTPException: Caso a cidade não seja encontrada.
 
     Returns:
-        Optional[CidadesSchema]: Representação da cidade encontrada ou None se não encontrada.
+        Optional[CidadeCreate]: Representação da cidade encontrada ou None se não encontrada.
     """
     # Busca pela cidade no banco de dados (ignora case com ilike)
     cidade = session.query(Cidade).filter(Cidade.__table__.c.cid_uf.ilike(f"%{uf.upper()}%")).first()
@@ -216,7 +264,7 @@ def buscar_cidade_uf(retorna_exception: bool, uf: str, session: Session) -> Opti
         if retorna_exception:
             raise HTTPException(status_code=404, detail=f"Cidade com UF '{uf}' não encontrada.")
 
-    return CidadesSchema(**cidade.as_dict()) if cidade else None
+    return CidadeOut(**cidade.as_dict()) if cidade else None
 
 def consiste_uf (uf: str) -> bool:
     """Verifica se a UF é válida.
